Further project reorganization
