--- project-template/src/lib-app/include/GameTitle.hpp
@@ -1,4 +1,0 @@
-#pragma once
-
-/// This value is populated by cmake variable from settings.cmake
-const std::string GAME_TITLE = "Example";
