--- project-template/src/lib-app/include/core/Settings.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-struct Settings
-{
-	float soundVolume = 50.f;
-	float musicVolume = 50.f;
-
-	std::string resourcesDir = "../resources";
-	bool skipMainMenu = false;
-};
