+++ project-template/src/lib-engines/src/engine/AudioEngine.cpp
@@ -1,3 +1,3 @@
-#include "AudioEngine.hpp"
+#include "engine/AudioEngine.hpp"
 
 void AudioEngine::update(const dgm::Time& time) {}
