+++ project-template/src/lib-engines/src/engine/GameRulesEngine.cpp
@@ -1,3 +1,3 @@
-#include "GameRulesEngine.hpp"
+#include "engine/GameRulesEngine.hpp"
 
 void GameRulesEngine::update(const dgm::Time& time) {}
