+++ project-template/src/lib-engines/src/engine/PhysicsEngine.cpp
@@ -1,3 +1,3 @@
-#include "PhysicsEngine.hpp"
+#include "engine/PhysicsEngine.hpp"
 
 void PhysicsEngine::update(const dgm::Time& time) {}
