+++ project-template/src/lib-engines/src/engine/RenderingEngine.cpp
@@ -1,4 +1,4 @@
-#include "RenderingEngine.hpp"
+#include "engine/RenderingEngine.hpp"
 
 void RenderingEngine::update(const dgm::Time& time) {}
