+++ project-template/src/lib-engines/src/render/FpsCounter.cpp
@@ -1,4 +1,4 @@
-#include "game/FpsCounter.hpp"
+#include "render/FpsCounter.hpp"
 
 void FpsCounter::update(const float dt)
 {
