+++ project-template/src/lib-settings/include/Settings.hpp
@@ -1,6 +1,11 @@
 #pragma once
 
+#include <filesystem>
+
 struct Settings
 {
-	
+    bool skipMainMenu = false;
+    std::filesystem::path resourcesDir = "../resources";
+    float soundVolume = 50.f;
+    float musicVolume = 50.f;
 };
